+++ packages/entities/entities-plugins/src/components/PluginEntityForm.vue
@@ -20,6 +20,7 @@
         :on-form-change="handleFreeFormUpdate"
         :on-model-updated="onModelUpdated"
         :on-validity-change="onValidityChange"
+        :plugin-name="formModel.name"
         :render-rules="PLUGIN_METADATA[formModel.name]?.freeformRenderRules"
         :schema="freeformSchema"
         @global-action="(name: GlobalAction, payload: any) => $emit('globalAction', name, payload)"