+++ packages/entities/entities-plugins/src/components/free-form/shared/layout/StandardLayout.vue
@@ -17,6 +17,14 @@
       >
         <RedisSelector />
       </FieldRenderer>
+
+      <!-- Identity Realms field (key-auth plugin only) -->
+      <FieldRenderer
+        v-slot="props"
+        :match="({ path }) => pluginName === 'key-auth' && path === 'config.identity_realms'"
+      >
+        <IdentityRealmsField v-bind="props" />
+      </FieldRenderer>
     </template>
 
     <template v-if="editorMode === 'form'">
@@ -149,6 +157,7 @@ export type Props<T extends FreeFormPluginData = any> = {
   /** FreeForm configuration */
   formConfig?: FormConfig<T>
   renderRules?: RenderRules
+  pluginName: string
 }
 </script>
 
@@ -169,6 +178,7 @@ import FieldRenderer from '../FieldRenderer.vue'
 import { REDIS_PARTIAL_INFO } from '../const'
 import RedisSelector from '../RedisSelector.vue'
 import { FIELD_RENDERERS } from '../composables'
+import IdentityRealmsField from '../../../fields/key-auth-identity-realms/FreeFormAdapter.vue'
 
 const { t } = createI18n<typeof english>('en-us', english)